QPID-8144: [Broker-J] Correct arithmetic type used when computing utilised direct memory (from int to long).
Cherry picked from 6.1.x b35dda09a8c0a3c99fd522518e899dfd5f53bd58

+++ common/src/main/java/org/apache/qpid/bytebuffer/QpidByteBuffer.java
@@ -722,7 +722,7 @@ public final class QpidByteBuffer
 
     public static long getAllocatedDirectMemorySize()
     {
-        return _pooledBufferSize * getNumberOfActivePooledBuffers();
+        return (long) _pooledBufferSize * getNumberOfActivePooledBuffers();
     }
 
     public static int getNumberOfActivePooledBuffers()
